Allow each shipyard to get assigned exactly once per distance a mining ship is away from it
Each shipyard is added to the list of positions that get assigned to mining ships once for every distance a mining ships is away from it. While the distance is not always exactly the travel time because of obstacles in the way, it should be better than just assigning the shipyard multiple times without checking the distance like we did before.